refactor(desktop): split render.rs into render/footer.rs and render/prompt.rs
Monolith split — render.rs was over the 500-line cap. Footer and prompt
rendering extracted to sibling modules under apps/desktop/src/render/.
